adding delete comment route

+++ src/services/delete-comment.ts
@@ -1,4 +1,5 @@
 import { CommentRepository } from "@/repositories/comment-repository";
+import { ResourceNotFoundError } from "./errors/resource-not-found-error";
 
 interface DeleteCommentServiceRequest {
   commentId: string;
@@ -8,6 +9,11 @@ export class DeleteComment {
   constructor(private commentRepository: CommentRepository) {}
 
   async execute({ commentId }: DeleteCommentServiceRequest) {
+    const findComment = await this.commentRepository.findById(commentId);
+    if (!findComment) {
+      throw new ResourceNotFoundError();
+    }
+
     await this.commentRepository.delete(commentId);
   }
 }
